Created Plate Counter, Stove Counter

+++ Assets/Scripts/CuttingCounter.cs
@@ -5,14 +5,21 @@ using UnityEngine;
 public class CuttingCounter : BaseCounter
 {
     // Start is called before the first frame update
-    [SerializeField] SO_KitchenObjects cutKitchenObject;
+    private const string CUT = "Cut";
+    [SerializeField] SO_CuttingRecipe[] cutKitchenObject;
+    [SerializeField] Animator cutanim;
+
+    private int cuttingcount;
      public override void Interact(PlayerMovement player)
     {
         if(!HasKitchenObject())
         {
-            if(player.HasKitchenObject())
+            if(player.HasKitchenObject())// i.e if player is carrying any ingredient 
             {
-                player.GetKithenObjects().SetKitchenObjectParent(this);
+                if(isCuttable(player.GetKithenObjects().GetScriptableObject())) //i.e if player is carrying that can be sliced
+                {
+                    player.GetKithenObjects().SetKitchenObjectParent(this);
+                }
             }
         }
         else
@@ -21,17 +28,55 @@ public class CuttingCounter : BaseCounter
             {
                 GetKithenObjects().SetKitchenObjectParent(player);
             }
+            else if(player.HasKitchenObject())
+            {
+                if(player.GetKithenObjects() is PlateKitchenObject)
+                {
+                    
+                    PlateKitchenObject plateKitchenObject = player.GetKithenObjects() as PlateKitchenObject;
+                    if(plateKitchenObject.TryAddIngredient(GetKithenObjects().GetScriptableObject()))
+                    {
+                        GetKithenObjects().DestroySelf();
+                    }
+                }
+            }
         }
     }
 
     public override void InteractAlternate(PlayerMovement Player)
     {
-        if(HasKitchenObject())
+        if(HasKitchenObject() && isCuttable(GetKithenObjects().GetScriptableObject()))
         {
+            //Play Cutting animation
+            if(cutanim!=null)cutanim.SetTrigger(CUT);
+
+            SO_KitchenObjects outputobj = GetOutputForInput(GetKithenObjects().GetScriptableObject());
             GetKithenObjects().DestroySelf();
             
-            Transform spawnItem = Instantiate(cutKitchenObject.prefab);
+            Transform spawnItem = Instantiate(outputobj.prefab);
             spawnItem.GetComponent<KithenObjects>().SetKitchenObjectParent(this);  
         }
+    }
+    private SO_KitchenObjects GetOutputForInput(SO_KitchenObjects inputKitchenObjectSO)
+    {
+        foreach(SO_CuttingRecipe CR in cutKitchenObject)
+        {
+            if(CR.input == inputKitchenObjectSO)
+            {
+                return CR.output;
+            }
+        }
+        return null;
+    }
+    private bool isCuttable(SO_KitchenObjects inputKitchenObjectSO)
+    {
+        foreach(SO_CuttingRecipe CR in cutKitchenObject)
+        {
+            if(CR.input == inputKitchenObjectSO)
+            {
+                return true;
+            }
+        }
+        return false;
     }
 }
